requirements.txt
requests~=2.31.0
requests_kerberos~=0.14.0
python-telegram-bot~=20.7
prettytable~=3.9.0
msrest~=0.7.1
aiohttp~=3.8.3
nextcord~=2.6.0